requirements.txt
PyQt5>=5.15.0
paho-mqtt>=1.6.0
